app/requirements.txt

beautifulsoup4==4.10.0
bs4==0.0.1
matplotlib==3.5.1
matplotlib-inline
nltk==3.6.7
numpy==1.22.1
pandas==1.4.0
pandas-datareader==0.10.0
pandocfilters
requests 
st-annotated-text==3.0.0
streamlit==1.5

